Use BigDecimal, fix bugs
This commit will make the program use BigDecimal, instead
of the double primitive. Unfortunately, BigDecimal doesn't
support exponents with decimal power, so I was forced to
use a hack, as I didn't want to add external dependencies.
Should work good enough. Also fixed two bugs: first, where
dot in decimal number was skipped because the line was
valueString += expr[++i] instead of valueString += expr[i++].
Second was the bug in PrefixNotation, which made the expression
evaluate in a wrong order.

+++ Notation.java
@@ -1,4 +1,5 @@
 import java.util.ArrayDeque;
+import java.math.BigDecimal;
 
 abstract class Notation {
 
@@ -70,6 +71,6 @@ abstract class Notation {
 	throw new WrongExpressionException(string + ": Unknown notation format.");
     }
 
-    abstract Tuple<ArrayDeque<Character>, ArrayDeque<Double>> tokenize(String expr) throws WrongExpressionException;
-    abstract double evaluate(Tuple<ArrayDeque<Character>,ArrayDeque<Double>> tokens) throws WrongExpressionException;
+    abstract Tuple<ArrayDeque<Character>, ArrayDeque<BigDecimal>> tokenize(String expr) throws WrongExpressionException;
+    abstract BigDecimal evaluate(Tuple<ArrayDeque<Character>,ArrayDeque<BigDecimal>> tokens) throws WrongExpressionException;
 }
